+++ CHANGELOG.md
@@ -55,6 +55,8 @@
     * Implement some more `zizmor` recommendations (\#3121, \#3122, \#3124, \#3125).
     * Remove `GitHubSecurityLab/actions-permissions/monitor` to avoid TLS issues (\#3124).
     * Bump `uv` version in GitHub Actions (\#3124).
+* Runner
+    * Added support for S3
 
 # 2.18.5